refactor!: Cleanup legacy simulator stuff
