+++ lib/simulator.js
@@ -1,6 +1,3 @@
-import SimulatorXcode6 from './simulator-xcode-6';
-import SimulatorXcode7 from './simulator-xcode-7';
-import SimulatorXcode73 from './simulator-xcode-7.3';
 import SimulatorXcode8 from './simulator-xcode-8';
 import SimulatorXcode9 from './simulator-xcode-9';
 import SimulatorXcode93 from './simulator-xcode-9.3';
@@ -12,13 +9,16 @@ import { getSimulatorInfo } from './utils';
 import xcode from 'appium-xcode';
 import { log, setLoggingPlatform } from './logger';
 
+const MIN_SUPPORTED_XCODE_VERSION = 8;
 
 function handleUnsupportedXcode (xcodeVersion) {
-  if (xcodeVersion.major < 6) {
-    throw new Error(`Tried to use an iOS simulator with xcode ` +
-                    `version ${xcodeVersion.versionString} but only Xcode version ` +
-                    `6.0.0 and up are supported`);
+  if (xcodeVersion.major < MIN_SUPPORTED_XCODE_VERSION) {
+    throw new Error(
+      `Tried to use an iOS simulator with xcode version ${xcodeVersion.versionString} but only Xcode version ` +
+      `${MIN_SUPPORTED_XCODE_VERSION} and up are supported`
+    );
   }
+  return xcodeVersion;
 }
 
 /**
@@ -48,7 +48,7 @@ async function getSimulator (udid, opts = {}) {
     devicesSetPath,
   } = opts;
 
-  const xcodeVersion = await xcode.getVersion(true);
+  const xcodeVersion = handleUnsupportedXcode(await xcode.getVersion(true));
   if (checkExistence) {
     const simulatorInfo = await getSimulatorInfo(udid, {
       devicesSetPath
@@ -64,24 +64,11 @@ async function getSimulator (udid, opts = {}) {
   // make sure we have the right logging prefix
   setLoggingPlatform(platform);
 
-  log.info(`Constructing ${platform} simulator for Xcode version ${xcodeVersion.versionString} ` +
-           `with udid '${udid}'`);
+  log.info(
+    `Constructing ${platform} simulator for Xcode version ${xcodeVersion.versionString} with udid '${udid}'`
+  );
   let SimClass;
   switch (xcodeVersion.major) {
-    // Early, unsupported Xcodes
-    case 1:
-    case 2:
-    case 3:
-    case 4:
-    case 5:
-      handleUnsupportedXcode(xcodeVersion);
-      break;
-    case 6:
-      SimClass = SimulatorXcode6;
-      break;
-    case 7:
-      SimClass = xcodeVersion.minor < 3 ? SimulatorXcode7 : SimulatorXcode73;
-      break;
     case 8:
       SimClass = SimulatorXcode8;
       break;
@@ -111,34 +98,4 @@ async function getSimulator (udid, opts = {}) {
   return result;
 }
 
-/**
- * Takes a set of options and finds the correct device string in order for Instruments to
- * identify the correct simulator.
- *
- * @param {object} opts - The options available are:
- *   - `deviceName` - a name for the device. If the given device name starts with `=`, the name, less the equals sign, is returned.
- *   - `platformVersion` - the version of iOS to use. Defaults to the current Xcode's maximum SDK version.
- *   - `forceIphone` - force the configuration of the device string to iPhone. Defaults to `false`.
- *   - `forceIpad` - force the configuration of the device string to iPad. Defaults to `false`.
- *   If both `forceIphone` and `forceIpad` are true, the device will be forced to iPhone.
- *
- * @return {string} The found device string, for example:
- *   'iPhone 5 (8.4)' with Xcode 7+
- *   'iPhone 5 (8.4 Simulator)' with Xcode 6+
- */
-async function getDeviceString (opts) {
-  let xcodeVersion = await xcode.getVersion(true);
-
-  handleUnsupportedXcode(xcodeVersion);
-
-  log.info(`Retrieving device name string for Xcode version ${xcodeVersion.versionString}`);
-  if (xcodeVersion.major >= 8) {
-    return await SimulatorXcode7.getDeviceString(opts);
-  } else if (xcodeVersion.major === 7) {
-    return await SimulatorXcode7.getDeviceString(opts);
-  } else if (xcodeVersion.major === 6) {
-    return await SimulatorXcode6.getDeviceString(opts);
-  }
-}
-
-export { getSimulator, getDeviceString };
+export { getSimulator };
