+++ lib/utils.js
@@ -4,10 +4,7 @@ import { exec } from 'teen_process';
 import { waitForCondition } from 'asyncbox';
 import { getVersion } from 'appium-xcode';
 import Simctl from 'node-simctl';
-import { fs, tempDir } from '@appium/support';
-import { Certificate } from './certificate';
-import path from 'path';
-import Simulator from './simulator-xcode-6';
+import { fs } from '@appium/support';
 
 
 const DEFAULT_SIM_SHUTDOWN_TIMEOUT = 30000;
@@ -49,7 +46,6 @@ function toBiometricDomainComponent (name) {
 // 1       No processes were matched.
 // 2       Invalid options were specified on the command line.
 // 3       An internal error occurred.
-
 async function pkill (appName, forceKill = false) {
   let args = forceKill ? ['-9'] : [];
   args.push('-x', appName);
@@ -204,108 +200,6 @@ async function safeRimRaf (delPath, tryNum = 0) {
   }
 }
 
-/**
- * Install an SSL certificate to a device with given udid
- * @param {string} pemText SSL pem text
- * @param {string} udid Identifier of the Simulator
- */
-async function installSSLCert (pemText, udid) {
-  // Check that openssl is installed on the path
-  try {
-    await fs.which('openssl');
-  } catch (e) {
-    log.debug(`customSSLCert requires openssl to be available on path`);
-    log.errorAndThrow(`Command 'openssl' not found`);
-  }
-
-  // Check that sqlite3 is installed on the path
-  try {
-    await fs.which('sqlite3');
-  } catch (e) {
-    log.debug(`customSSLCert requires sqlite3 to be available on path`);
-    log.errorAndThrow(`Command 'sqlite3' not found`);
-  }
-
-  // Create a temporary file to store PEM text
-  // (a temp file is necessary to run `openssl` shell commands, can't be done in memory)
-  let tempFileName = path.resolve(await tempDir.openDir(), 'temp-ssl-cert.pem');
-  let pathToKeychain = new Simulator(udid).getDir();
-  await fs.writeFile(tempFileName, pemText);
-  try {
-    await fs.stat(pathToKeychain);
-  } catch (e) {
-    log.debug(`Could not install SSL certificate. No simulator with udid '${udid}'`);
-    log.errorAndThrow(e);
-  }
-
-  // Do the certificate installation
-  let certificate = new Certificate(tempFileName);
-  log.debug(`Installing certificate to ${pathToKeychain}`);
-  await certificate.add(pathToKeychain);
-
-  // Remove the temporary file
-  await fs.unlink(tempFileName);
-
-  return certificate;
-}
-
-async function uninstallSSLCert (pemText, udid) {
-  try {
-    let tempFileName = path.resolve(__dirname, 'temp-ssl-cert.pem');
-    let pathToKeychain = path.resolve(new Simulator(udid).getDir());
-    await fs.writeFile(tempFileName, pemText);
-    let certificate = new Certificate(tempFileName);
-    await certificate.remove(pathToKeychain);
-    await fs.unlink(tempFileName);
-    return certificate;
-  } catch (e) {
-    log.debug(`Could not uninstall SSL certificate. No simulator with udid '${udid}'`);
-    log.errorAndThrow(e);
-  }
-}
-
-/**
- * Check if the Simulator already has this SSL certificate
- * @param {string} pemText PEM text of SSL cert
- * @param {string} udid Identifier of the Simulator
- */
-async function hasSSLCert (pemText, udid) {
-  const tempFileName = path.resolve(await tempDir.openDir(), 'temp-ssl-cert.pem');
-  const pathToKeychain = new Simulator(udid).getDir();
-  await fs.writeFile(tempFileName, pemText);
-  const certificate = new Certificate(tempFileName);
-  return certificate.has(pathToKeychain);
-}
-
-/**
- * Runs a command line sqlite3 query
- *
- * @param {string} db - Full path to sqlite database
- * @param {string} query - The actual query string
- * @param {...string} queryParams - The list of query parameters
- * @returns {string} sqlite command stdout
- */
-async function execSQLiteQuery (db, query, ...queryParams) {
-  query = query.replace(/\n+/g, ' ');
-  let queryTokens = query.split('?');
-  let formattedQuery = [];
-  queryParams
-    .map((param) => `${param}`)
-    .forEach((param, i) => {
-      formattedQuery.push(queryTokens[i]);
-      formattedQuery.push(param.replace(/'/g, "''"));
-    });
-  formattedQuery.push(queryTokens[queryTokens.length - 1]);
-
-  log.debug(`Executing SQL query "${formattedQuery.join('')}" on '${db}'`);
-  try {
-    return (await exec('sqlite3', ['-line', db, formattedQuery.join('')])).stdout;
-  } catch (err) {
-    throw new Error(`Cannot execute SQLite query "${formattedQuery.join('')}" to '${db}'. ` +
-      `Original error: ${err.stderr}`);
-  }
-}
-
 async function getDeveloperRoot () {
   const {stdout} = await exec('xcode-select', ['-p']);
   return stdout.trim();
@@ -342,8 +236,8 @@ async function launchApp (simctl, bundleId, timeoutMs = 25000) {
   try {
     await waitForCondition(async () => {
       try {
-        const stdout = await simctl.launchApp(MOBILE_SAFARI_BUNDLE_ID, 1);
-        return PROCESS_LAUNCH_OK_PATTERN(MOBILE_SAFARI_BUNDLE_ID).test(stdout);
+        const stdout = await simctl.launchApp(bundleId, 1);
+        return PROCESS_LAUNCH_OK_PATTERN(bundleId).test(stdout);
       } catch (err) {
         lastError = err.stderr || err.message;
       }
@@ -364,10 +258,6 @@ export {
   safeRimRaf,
   simExists,
   getSimulatorInfo,
-  installSSLCert,
-  uninstallSSLCert,
-  hasSSLCert,
-  execSQLiteQuery,
   toBiometricDomainComponent,
   getDeveloperRoot,
   activateApp,
