+++ test/unit/simulator-specs.js
@@ -1,6 +1,6 @@
 // transpile:mocha
 
-import { getSimulator, getDeviceString } from '../..';
+import { getSimulator } from '../..';
 import Simctl from 'node-simctl';
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
@@ -8,8 +8,6 @@ import sinon from 'sinon';
 import { devices } from '../assets/deviceList';
 import B from 'bluebird';
 import xcode from 'appium-xcode';
-import SimulatorXcode6 from '../../lib/simulator-xcode-6';
-import SimulatorXcode7 from '../../lib/simulator-xcode-7';
 import SimulatorXcode8 from '../../lib/simulator-xcode-8';
 import SimulatorXcode9 from '../../lib/simulator-xcode-9';
 import SimulatorXcode93 from '../../lib/simulator-xcode-9.3';
@@ -21,7 +19,7 @@ import SimulatorXcode11_4 from '../../lib/simulator-xcode-11.4';
 chai.should();
 chai.use(chaiAsPromised);
 
-const UDID = devices['7.1'][0].udid;
+const UDID = devices['8.1'][0].udid;
 
 describe('simulator', function () {
   let xcodeMock;
@@ -39,17 +37,15 @@ describe('simulator', function () {
 
   describe('getSimulator', function () {
     it('should create a simulator with default xcode version', async function () {
-      let xcodeVersion = {major: 6, versionString: '6.0.0'};
+      let xcodeVersion = {major: 8, versionString: '8.0.0'};
       xcodeMock.expects('getVersion').returns(B.resolve(xcodeVersion));
 
       let sim = await getSimulator(UDID);
       sim.xcodeVersion.should.equal(xcodeVersion);
-      sim.constructor.name.should.be.eql(SimulatorXcode6.name);
+      sim.constructor.name.should.be.eql(SimulatorXcode8.name);
     });
 
     const xcodeVersions = [
-      [6, 0, '6.0.0', SimulatorXcode6],
-      [7, 0, '7.0.0', SimulatorXcode7],
       [8, 0, '8.0.0', SimulatorXcode8],
       [9, 0, '9.0.0', SimulatorXcode9],
       [9, 3, '9.3.0', SimulatorXcode93],
@@ -79,7 +75,7 @@ describe('simulator', function () {
     });
 
     it('should list stats for sim', async function () {
-      let xcodeVersion = {major: 6, versionString: '6.0.0'};
+      let xcodeVersion = {major: 8, versionString: '8.0.0'};
       xcodeMock.expects('getVersion').atLeast(1).returns(B.resolve(xcodeVersion));
 
       let sims = [
@@ -98,108 +94,6 @@ describe('simulator', function () {
       stats[0].name.should.equal('Resizable iPhone');
       stats[1].state.should.equal('Shutdown');
       stats[1].name.should.equal('Resizable iPad');
-    });
-  });
-
-
-  describe('getDeviceString', function () {
-    describe('Xcode 6', function () {
-      let xcodeVersion = {major: 6, versionString: '6.0.0'};
-
-      beforeEach(function () {
-        xcodeMock.expects('getVersion').returns(B.resolve(xcodeVersion));
-      });
-
-      it('should get the correct device for iOS 8+', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(8.4));
-        let device = await getDeviceString();
-        device.should.equal('iPhone 6 (8.4 Simulator)');
-      });
-      it('should get the correct device for iOS 8+ when platform version passed in', async function () {
-        let device = await getDeviceString({platformVersion: '8.1'});
-        device.should.equal('iPhone 6 (8.1 Simulator)');
-      });
-      it('should get the correct device for iOS 7+', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(7.1));
-        let device = await getDeviceString();
-        device.should.equal('iPhone 5s (7.1 Simulator)');
-      });
-      it('should get the correct device for iOS 7+ when platform version passed in', async function () {
-        let device = await getDeviceString({platformVersion: '7.1'});
-        device.should.equal('iPhone 5s (7.1 Simulator)');
-      });
-      it('should pass through device name when passed with =', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(8.4));
-        let device = await getDeviceString({deviceName: '=fancy device'});
-        device.should.equal('fancy device');
-      });
-      it('should add a device name when passed without =', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(8.4));
-        let device = await getDeviceString({deviceName: 'fancy device'});
-        device.should.equal('fancy device (8.4 Simulator)');
-      });
-      it('should handle string platform version', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve('8.4'));
-        let device = await getDeviceString();
-        device.should.equal('iPhone 6 (8.4 Simulator)');
-      });
-      it('should strip " Simulator" when not necessary', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve('8.4'));
-        let device = await getDeviceString({deviceName: 'iPhone 6 Simulator'});
-        device.should.equal('iPhone 6 (8.4 Simulator)');
-      });
-      it('should handle bare "iPhone"', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve('8.4'));
-        let device = await getDeviceString({deviceName: 'iPhone'});
-        device.should.equal('iPhone 6 (8.4 Simulator)');
-      });
-    });
-
-    describe('Xcode 7', function () {
-      let xcodeVersion = {major: 7, versionString: '7.0.0'};
-
-      beforeEach(function () {
-        xcodeMock.expects('getVersion').returns(B.resolve(xcodeVersion));
-      });
-
-      it('should get the correct device for iOS 8+', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(8.4));
-        let device = await getDeviceString();
-        device.should.equal('iPhone 6 (8.4)');
-      });
-      it('should get the correct device for iOS 8+ when platform version passed in', async function () {
-        let device = await SimulatorXcode7.getDeviceString({platformVersion: '8.1'});
-        device.should.equal('iPhone 6 (8.1)');
-      });
-      it('should get the correct device for iOS 9+', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(9));
-        let device = await getDeviceString();
-        device.should.equal('iPhone 6 (9.0) [');
-      });
-      it('should get the correct device for iOS 9+ when platform version passed in', async function () {
-        let device = await getDeviceString({platformVersion: '9.0'});
-        device.should.equal('iPhone 6 (9.0) [');
-      });
-      it('should pass through device name when passed with =', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(9.0));
-        let device = await getDeviceString({deviceName: '=fancy device'});
-        device.should.equal('fancy device');
-      });
-      it('should add a device name when passed without =', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve(9.0));
-        let device = await getDeviceString({deviceName: 'fancy device'});
-        device.should.equal('fancy device (9.0)');
-      });
-      it('should strip " Simulator" when not necessary', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve('9.0'));
-        let device = await getDeviceString({deviceName: 'iPhone 6 Simulator'});
-        device.should.equal('iPhone 6 (9.0) [');
-      });
-      it('should handle bare "iPhone"', async function () {
-        xcodeMock.expects('getMaxIOSSDK').returns(B.resolve('9.0'));
-        let device = await getDeviceString({deviceName: 'iPhone'});
-        device.should.equal('iPhone 6 (9.0) [');
-      });
     });
   });
 });
